Refactor wallet settings module
